+++ app/routes/application.js
@@ -15,7 +15,7 @@ function createNewApplication (req) {
       choices: {},
       references: {},
       candidate: {},
-      'contact-details': {},
+      'contact-information': {},
       'additional-support': {},
       safeguarding: {},
       degree: {},
@@ -141,7 +141,7 @@ module.exports = router => {
 
   require('./application/choices')(router)
   require('./application/personal-information')(router)
-  require('./application/contact-details')(router)
+  require('./application/contact-information')(router)
   require('./application/english-language')(router)
   require('./application/work-history')(router)
   require('./application/school-experience')(router)
