+++ app/utils/index.js
@@ -136,7 +136,7 @@ const hasCompletedApplication = req => {
     module.exports.hasCompletedSection(application.choices) &&
     module.exports.hasCompletedSection(application.references) &&
     module.exports.hasCompletedSection(application.candidate) &&
-    module.exports.hasCompletedSection(application['contact-details']) &&
+    module.exports.hasCompletedSection(application['contact-information']) &&
     module.exports.hasCompletedSection(application['additional-support']) &&
     module.exports.hasCompletedSection(application['work-history']) &&
     module.exports.hasCompletedSection(application['school-experience']) &&
