+++ app/views/_includes/update-contact-details.njk
@@ -1,8 +1,8 @@
 {% if not data.flags.self_amend_contact_details %}
   {{ appRelated({
-    title: "Updating your contact details",
+    title: "Updating your contact information",
     content: {
-      html: "<p class=\"govuk-body-s\">You can update your contact details at any point during the application process. To do this, email us at <a href=\"mailto:becomingateacher@digital.education.gov.uk\">becomingateacher<wbr>@digital.education.gov.uk</a>.</p>"
+      html: "<p class=\"govuk-body-s\">You can update your contact information at any point during the application process. To do this, email us at <a href=\"mailto:becomingateacher@digital.education.gov.uk\">becomingateacher<wbr>@digital.education.gov.uk</a>.</p>"
     }
   }) }}
 {% endif %}
