+++ app/views/application/contact-information/uk-address.njk
@@ -1,7 +1,7 @@
 {% extends "_form.njk" %}
 
 {% set title = "What is your address?" %}
-{% set international = applicationValue(["contact-details", "address-type"]) == "international" %}
+{% set international = applicationValue(["contact-information", "address-type"]) == "international" %}
 
 {% block pageNavigation %}
   {% if referrer %}
@@ -10,7 +10,7 @@
     }) }}
   {% else %}
     {{ govukBackLink({
-      href: "/application/" + applicationId + "/contact-details/where-do-you-live"
+      href: "/application/" + applicationId + "/contact-information/where-do-you-live"
     }) }}
   {% endif %}
 {% endblock %}
@@ -22,21 +22,21 @@
         text: "Enter your address"
       },
       autocomplete: "street-address"
-    } | decorateApplicationAttributes(["contact-details", "address", "street-address"])) }}
+    } | decorateApplicationAttributes(["contact-information", "address", "street-address"])) }}
   {% else %}
     {{ govukInput({
       label: {
         html: "Building and street <span class=\"govuk-visually-hidden\">line 1 of 2</span>"
       },
       autocomplete: "address-line1"
-    } | decorateApplicationAttributes(["contact-details", "address", "line1"])) }}
+    } | decorateApplicationAttributes(["contact-information", "address", "line1"])) }}
 
     {{ govukInput({
       label: {
         html: "<span class=\"govuk-visually-hidden\">Building and street line 2 of 2</span>"
       },
       autocomplete: "address-line2"
-    } | decorateApplicationAttributes(["contact-details", "address", "line2"])) }}
+    } | decorateApplicationAttributes(["contact-information", "address", "line2"])) }}
 
     {{ govukInput({
       label: {
@@ -44,7 +44,7 @@
       },
       classes: "govuk-!-width-two-thirds",
       autocomplete: "address-level2"
-    } | decorateApplicationAttributes(["contact-details", "address", "level2"])) }}
+    } | decorateApplicationAttributes(["contact-information", "address", "level2"])) }}
 
     {{ govukInput({
       label: {
@@ -52,7 +52,7 @@
       },
       classes: "govuk-!-width-two-thirds",
       autocomplete: "address-level1"
-    } | decorateApplicationAttributes(["contact-details", "address", "level1"])) }}
+    } | decorateApplicationAttributes(["contact-information", "address", "level1"])) }}
 
     {{ govukInput({
       label: {
@@ -60,7 +60,7 @@
       },
       classes: "govuk-input--width-10",
       autocomplete: "postal-code"
-    } | decorateApplicationAttributes(["contact-details", "address", "postal-code"])) }}
+    } | decorateApplicationAttributes(["contact-information", "address", "postal-code"])) }}
   {% endif %}
 
   {{ govukButton({
