+++ app/views/application/review.njk
@@ -54,7 +54,7 @@
     {% include "_includes/review/personal-information.njk" %}
 
     <h3 class="govuk-heading-m">Contact information</h3>
-    {% include "_includes/review/contact-details.njk" %}
+    {% include "_includes/review/contact-information.njk" %}
   </section>
 
   <section class="app-section">
