+++ app/views/application/submitted.njk
@@ -56,8 +56,8 @@
     {% include "_includes/review/personal-information.njk" %}
 
     <h3 class="govuk-heading-m">Contact information</h3>
-    {% set contactDetails = applicationValue(["contact-details"]) %}
-    {% include "_includes/review/contact-details.njk" %}
+    {% set contactDetails = applicationValue(["contact-information"]) %}
+    {% include "_includes/review/contact-information.njk" %}
   </section>
 
   <section class="govuk-!-margin-bottom-8">
