+++ app/views/dashboard/index.njk
@@ -36,8 +36,8 @@
 
   {% include "dashboard/_course-choices.njk" %}
 
-  <h2 class="govuk-heading-m">Updating your contact details</h2>
-  <p class="govuk-body">To update your contact details, email us at <a href="mailto:becomingateacher@digital.education.gov.uk">becomingateacher<wbr>@digital.education.gov.uk</a>.</p>
+  <h2 class="govuk-heading-m">Updating your contact information</h2>
+  <p class="govuk-body">To update your contact information, email us at <a href="mailto:becomingateacher@digital.education.gov.uk">becomingateacher<wbr>@digital.education.gov.uk</a>.</p>
 
   {% include "dashboard/_previous-applications.njk" %}
 {% endblock %}
